fix(playground-server): use chat-completions openai transform to reduce duplicate files
