+++ sites/playground/server/tsconfig.dev.json
@@ -11,6 +11,9 @@
       ],
       "@opentiny/genui-sdk-materials-angular-opentiny-ng/render-config": [
         "../../../packages/materials/angular-opentiny-ng/src/render-config/index.ts"
+      ],
+      "@opentiny/genui-sdk-chat-completions": [
+        "../../../packages/chat-completions/src/index.ts"
       ]
     }
   },
